+++ @xen-orchestra/fs/src/nfs.js
@@ -4,6 +4,8 @@ import { forEach } from 'lodash'
 
 import LocalHandler from './local'
 
+const DEFAULT_NFS_OPTIONS = 'vers=3'
+
 export default class NfsHandler extends LocalHandler {
   get type () {
     return 'nfs'
@@ -52,12 +54,12 @@ export default class NfsHandler extends LocalHandler {
 
   async _mount () {
     await fs.ensureDir(this._getRealPath())
-    const { host, path, port } = this._remote
+    const { host, path, port, options } = this._remote
     return execa('mount', [
       '-t',
       'nfs',
       '-o',
-      'vers=3',
+      DEFAULT_NFS_OPTIONS + (options !== undefined ? `,${options}` : ''),
       `${host}${port !== undefined ? ':' + port : ''}:${path}`,
       this._getRealPath(),
     ])